O365 phishlet
Tested and catches all logins from ADFS and O365. Depending on the local ADFS implementation this might need to be changed.

In addition, there is a workaround for the missing detection of the :443. Fields on the sub_filters & proxy_hosts need to be manually filled into the yaml before launch. By allowing to use params for those things (separate PR) this can be avoided in the future.