Add "security" TLD
Hey there,
After quite a long debugging session, I noticed that `.security` TLD wasn't part of the URL matching regex. My company's TLD is `.security`, and all my tests were failing when it came to ADFS and `auto_filter` feature.
If this could be merged so that all `.security` domain work with this tool, that would be awesome.

Thanks a lot for this awesome work.