Fix session table display bug
The session summary display table in terminal.go previously used
Sprintf for formatting text (phishlet, username, password, and URL).

When any of these contined a '%' followed by specific characters,
as might happen in a complex password, Sprintf would read the string
as having a placeholder and display a "(MISSING)" message in the text
since no placeholder was supplied.

This commit changes Sprintf to Sprint for values which should not
contain format placeholders. Since Sprint does not accept patterns,
there is no chance of error even if the text contains a '%'.